Autofill existing information, update information if existing.
When typing a Team Name or Team Number into the AddFragment, if it finds that the name or number already exists in the database, it will automatically fill out the rest of the data. Also, when adding to the database, if an entry of the same team number is found, it will update that entry instead of making a new one.